\name{Butterfly}
\alias{Butterfly}
\docType{data}
\title{Wingspans of butterflies in different sites}
\description{Imaginary data set of wingspans (in cm) of butterflies in different patches of different sites}
\usage{data("Butterfly")}
\format{
  A data frame with 300 observations of the following 4 variables.
  \describe{
    \item{\code{Sex}}{a character vector: M=males; F=Females}
    \item{\code{Site}}{a numeric vector: 3 different sites}
    \item{\code{Patch}}{a numeric vector: 5 patches per site}
    \item{\code{Wingspan}}{a numeric vector}
  }
}
\source{
Data created by Estelle Laurent
}
\references{
Data created by Estelle Laurent
}
\keyword{datasets}
