% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ParamANOVA-package.R
\docType{package}
\name{ParamANOVA-package}
\alias{ParamANOVA}
\alias{ParamANOVA-package}
\title{ParamANOVA: Fully Parameterizable ANOVA Tests}
\description{
Allows the user to perform ANOVA tests (in a strict sense: continuous and normally-distributed Y variable and 1 or more factorial/categorical X variable(s)), with the possibility to specify the type of sum of squares (1, 2 or 3), the types of variables (Fixed or Random) and their relationships (crossed or nested) with the sole function of the package (FullyParamANOVA()). The resulting outputs are the same as in 'SAS' software. A dataset (Butterfly) to test the function is also joined.
}
\author{
\strong{Maintainer}: Victor Brans \email{vicbrans003@gmail.com} (\href{https://orcid.org/0000-0002-8960-100X}{ORCID})

}
\keyword{internal}
